fix: move pkg
